backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
pymongo>=4.6
